% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district.R
\name{getDistrict}
\alias{getDistrict}
\title{Plot districts within a province}
\usage{
getDistrict(n_province = NULL, id = NULL)
}
\arguments{
\item{n_province}{The name of the province to plot. The notation should be respected.To get the notation: provinces()}

\item{id}{the id of the province.To get the id of each province provinces()}
}
\value{
return a sf object
}
\description{
Helps to plot the shape of districts within a province.
}
\examples{
\dontrun{
#Use DISTRICT name
prov=getDistrict("Casablanca")
plot(prov$coordinates)
#Use id
prov=getDistrict(id=141)
plot(prov$coordinates)
}
}
