% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district.R
\name{getMultiDistrict}
\alias{getMultiDistrict}
\title{Plot districts of multiple provinces}
\usage{
getMultiDistrict(n_province = NULL, id = NULL)
}
\arguments{
\item{n_province}{vector of The name of the province to plot. The notation should be respected.To get the notation: provinces()}

\item{id}{vector of provinces id.Each province has an id. To get the id of each province  : provinces()}
}
\value{
return a sf object
}
\description{
Helps to plot the shape of districts of multiple provinces.
}
\examples{
\dontrun{
prov=getMultiDistrict(c("Tanger-Assilah","Fahs-Anjra"))
plot(prov$coordinates)
prov=getMultiDistrict(id=c(227,511))
plot(prov$coordinates)
}
}
