% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/province.R
\name{getMultiProvince}
\alias{getMultiProvince}
\title{Plot multiple regions}
\usage{
getMultiProvince(n_region = NULL, id = NULL)
}
\arguments{
\item{n_region}{vector of the name of the regions to plot. The notation should be respected.To get the notation execute: regions()}

\item{id}{vector of regions id.Each region has an id. To get the id of each region please execute : regions()}
}
\value{
return a sf object
}
\description{
Helps to plot the shape of multiple regions.
}
\examples{
\dontrun{
# Plot two regions :Casablanca-Settat and Rabat-Sale-Kenitra
regions=getMultiProvince(c("Casablanca-Settat","Rabat-Sale-Kenitra"))
plot(regions$coordinates)
#Plot provinces of Oriental and Tanger-Tetouan-AL-Hoceima
regions=getMultiProvince(id=c(1,2))
plot(regions$coordinates)
}

}
