% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{getMultiRegion}
\alias{getMultiRegion}
\title{Plot multiple regions}
\usage{
getMultiRegion(name = NULL, id = NULL)
}
\arguments{
\item{name}{vector of the name of the regions to plot. The notation should be respected.To get the notation execute: regions()}

\item{id}{vector of regions id.Each region has an id. To get the id of each region please execute : regions()}
}
\value{
return a sf object
}
\description{
Helps to plot the shape of multiple regions.
}
\examples{
\dontrun{
#Plot two regions :Casablanca-Settat and Laayoune-Sakia-El-Hamra
regions=getMultiRegion(c("Eddakhla-Oued-Eddahab","Laayoune-Sakia-El-Hamra"))
plot(regions$coordinates)
#Plot the map of Morocco
mar=getMultiRegion(id=1:12)
plot(mar$coordinates)
}
}
