% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/province.R
\name{getPathRegion}
\alias{getPathRegion}
\title{Internal function}
\usage{
getPathRegion(n_region = NULL, id = NULL)
}
\arguments{
\item{n_region}{name}

\item{id}{id of region}
}
\description{
Internal function
}
