% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/province.R
\name{getProvince}
\alias{getProvince}
\title{Plot provinces within a region}
\usage{
getProvince(n_region = NULL, id = NULL)
}
\arguments{
\item{n_region}{The name of the region to plot. The notation should be respected.To get the notation : regions()}

\item{id}{Each region has an id. To get the id of each region : regions()}
}
\value{
return a sf object
}
\description{
Helps to plot the shape of provinces within a region.
}
\examples{
\dontrun{
#Use region name
region=getProvince("Tanger-Tetouan-Al-Hoceima")
plot(region$coordinates)
#Use id
region=getProvince(id=1) #Don't forget getRegion(id=1) not getRegion(1)
plot(region$coordinates)
}
}
