% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{getRegion}
\alias{getRegion}
\title{Plot regions}
\usage{
getRegion(name = NULL, id = NULL)
}
\arguments{
\item{name}{The name of the region to plot. The notation should be respected.To get the notation execute: regions()}

\item{id}{Each region has an id. To get the id of each region please execute : regions()}
}
\value{
return a sf object
}
\description{
plot the shape of each region
}
\examples{
\dontrun{

region=getRegion("Tanger-Tetouan-Al-Hoceima") #Use region name
#plot(region$coordinates)
region=getRegion(id=1) # use id
plot(region$coordinates)
}

}
