% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notation.R
\name{provinces}
\alias{provinces}
\title{Notation : provinces}
\usage{
provinces()
}
\value{
return a dataframe
}
\description{
Hepls to respect the notation and to get the id and the name of each province
}
\examples{
\dontrun{
province=provinces()
province
}
}
