% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notation.R
\name{regions}
\alias{regions}
\title{Notation : regions}
\usage{
regions()
}
\value{
return a dataframe
}
\description{
Hepls to respect the notation and to get the id and the name of each region
}
\examples{
\dontrun{
region=regions()
region}
}
